requirements.txt
flask
requests
yml
PyYAML
uptime
prometheus_flask_exporter
mysql-connector-python